+++ unit_tests/test_mer_dna.cc
@@ -304,11 +304,11 @@ TEST(MerDNASimple, MultipleSize) {
   mer_dna::k(10);
   mer_dna1::k(50);
   mer_dna2::k(100);
-  EXPECT_EQ(10, mer_dna::k());
+  EXPECT_EQ((unsigned int)10, mer_dna::k());
   EXPECT_EQ(0, mer_dna::class_index);
-  EXPECT_EQ(50, mer_dna1::k());
+  EXPECT_EQ((unsigned int)50, mer_dna1::k());
   EXPECT_EQ(1, mer_dna1::class_index);
-  EXPECT_EQ(100, mer_dna2::k());
+  EXPECT_EQ((unsigned int)100, mer_dna2::k());
   EXPECT_EQ(2, mer_dna2::class_index);
 }
